make all scripts use get_renders_root() function
